기대평 / 문의하기 삭제 시 확인창 뜨게 수정(강사님 지적사항)

+++ Soso/src/main/webapp/WEB-INF/views/group/faq/list.jsp
@@ -56,7 +56,7 @@
 									    	<pre name="content" id="content" readonly>${tmp.q_content}</pre>
 								        	<c:if test="${ empty tmp.a_answer && sessionScope.id == tmp.q_writer}">
 									        	<a href="${pageContext.request.contextPath}/group/faq/updateform?num=${tmp.num}" id="update">수정</a>
-									        	<a href="${pageContext.request.contextPath}/group/faq/delete?num=${tmp.num}&group_num=${tmp.group_num}" id="delete">삭제</a>
+									        	<a href="${pageContext.request.contextPath}/group/faq/delete?num=${tmp.num}&group_num=${tmp.group_num}" onclick="return confirm('삭제하시겠습니까?')" id="delete">삭제</a>
 								        	</c:if>
 										</div>
 										<div class="qna_answer">
@@ -73,7 +73,7 @@
 								        	</c:if>
 								        <c:if test="${not empty tmp.a_answer && dto.manager_id == sessionScope.id }">
 								        	<a href="${pageContext.request.contextPath}/group/answer/updateform?num=${tmp.num}" id="updateAnswer" >수정</a>
-								        	<a href="${pageContext.request.contextPath}/group/answer/delete?num=${tmp.num}&group_num=${tmp.group_num}" id="deleteAnswer">삭제</a>						        
+								        	<a href="${pageContext.request.contextPath}/group/answer/delete?num=${tmp.num}&group_num=${tmp.group_num}" onclick="return confirm('삭제하시겠습니까?')" id="deleteAnswer">삭제</a>						        
 								        </c:if>
 								       </div>
 								    </td>
